Refactor uint8_t to uint32_t

+++ tasmota/tasmota_xdrv_driver/xdrv_01_9_webserver.ino
@@ -3689,7 +3689,7 @@ void CmndCors(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv01(uint8_t function)
+bool Xdrv01(uint32_t function)
 {
   bool result = false;
