+++ tasmota/tasmota_xdrv_driver/xdrv_13_display.ino
@@ -2787,7 +2787,7 @@ void AddValue(uint8_t num,float fval) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv13(uint8_t function)
+bool Xdrv13(uint32_t function)
 {
   bool result = false;
