+++ tasmota/tasmota_xdrv_driver/xdrv_34_wemos_motor_v1.ino
@@ -268,7 +268,7 @@ void WMotorV2command(uint8_t *data, uint8_t len)    // process V2 request
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv34(uint8_t function)
+bool Xdrv34(uint32_t function)
 {
   if (!I2cEnabled(XI2C_44)) { return false; }
