+++ tasmota/tasmota_xdrv_driver/xdrv_81_esp32_webcam.ino
@@ -1422,7 +1422,7 @@ void WcStatsShow(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv81(uint8_t function) {
+bool Xdrv81(uint32_t function) {
   bool result = false;
 
   switch (function) {
