+++ tasmota/tasmota_xdrv_driver/xdrv_87_esp32_sonoff_tm1621.ino
@@ -566,7 +566,7 @@ void CmndDspSpeed(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv87(uint8_t function) {
+bool Xdrv87(uint32_t function) {
   bool result = false;
 
   if (FUNC_INIT == function) {
