+++ tasmota/tasmota_xnrg_energy/xnrg_06_pzem_dc.ino
@@ -147,7 +147,7 @@ bool PzemDcCommand(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xnrg06(uint8_t function)
+bool Xnrg06(uint32_t function)
 {
   bool result = false;
