+++ tasmota/tasmota_xnrg_energy/xnrg_09_dds2382.ino
@@ -110,7 +110,7 @@ void Dds2382DrvInit(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xnrg09(uint8_t function)
+bool Xnrg09(uint32_t function)
 {
   bool result = false;
