+++ tasmota/tasmota_xnrg_energy/xnrg_18_sdm72.ino
@@ -207,7 +207,7 @@ void Sdm72Show(bool json) {
  * Interface
 \*********************************************************************************************/
 
-bool Xnrg18(uint8_t function)
+bool Xnrg18(uint32_t function)
 {
   bool result = false;
