+++ tasmota/tasmota_xnrg_energy/xnrg_30_dummy.ino
@@ -159,7 +159,7 @@ void NrgDummyDrvInit(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xnrg30(uint8_t function) {
+bool Xnrg30(uint32_t function) {
   bool result = false;
 
   switch (function) {
